requirements.txt
flask==3.0.0
flask-cors==4.0.0
apscheduler==3.10.4
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
newspaper3k==0.2.8
scikit-learn==1.3.2
numpy==1.26.2
openai==1.3.0
lxml==4.9.3
feedparser==6.0.10
